+++ src/modules/swap/ModuleSwap.vue
@@ -19,7 +19,7 @@
             =====================================================================================
               From / Amount to Swap / To / Amount to Recieve
             =====================================================================================
-            -->
+          -->
           <v-row class="align-center justify-space-between">
             <v-col cols="12">
               <v-skeleton-loader v-if="isLoading" type="text" width="375px" />
@@ -72,179 +72,36 @@
           </v-row>
 
           <!--
-          =====================================================================================
-            User Message Block: store your Bitcoin on Ethereum
-          =====================================================================================
-          -->
-          <user-msg-block
-            v-if="notEnoughEth && !isLoading"
-            class="mt-5"
-            :message="msg.storeBitcoin"
-          >
-            <div class="mt-3 mx-n1">
-              <mew-button
-                btn-size="small"
-                btn-style="outline"
-                title="Swap Your Bitcoin to Ether"
-                class="ma-1"
-                :has-full-width="$vuetify.breakpoint.xsOnly"
-              />
-              <mew-button
-                btn-size="small"
-                btn-style="outline"
-                title="Buy Ether"
-                class="ma-1"
-                :has-full-width="$vuetify.breakpoint.xsOnly"
-                @click.native="buyEth"
-              />
-            </div>
-          </user-msg-block>
-
-          <!--
-          =====================================================================================
-            User Message Block: store your Bitcoin on Ethereum
-          =====================================================================================
-          -->
-          <user-msg-block
-            v-if="
-              toTokenType.value && toTokenType.value.toLowerCase() == 'bitcoin'
-            "
-            class="mt-5"
-            :message="msg.lowBalance"
-          >
-            <div class="border-top mt-3">
-              <v-expansion-panels
-                flat
-                class="expansion-panels--remove-paddings"
-              >
-                <v-expansion-panel>
-                  <v-expansion-panel-header
-                    color="tableHeader"
-                    class="textPrimaryModule--text"
-                  >
-                    How can I get wrapped Bitcoin?
-                  </v-expansion-panel-header>
-                  <v-expansion-panel-content color="tableHeader" class="pa-0">
-                    <div class="textPrimaryModule--text mb-2">
-                      When you swap to Bitcoin, it is moved to the Bitcoin
-                      blockchain, & requires a Bitcoin wallet. In order to keep
-                      Bitcoin in MyEtherWallet, you can swap to wrapped Bitcoin
-                      instead. Wrapped Bitcoin is an Ethereum token, with a
-                      value approximately equal to 1 BTC. Wrapped Bitcoins can
-                      be stored in MEW, and can be used as any other Ethereum
-                      asset: you can swap it to other tokens, use it as
-                      collateral in DeFi apps, etc. There are multiple kinds of
-                      wrapped Bitcoins, but they roughly do the same thing.
-                      <a
-                        href="https://kb.myetherwallet.com/en/swap/btc-to-ethereum/"
-                      >
-                        Learn more about Wrapped Bitcoin.
-                      </a>
-                    </div>
-                    <v-row class="mt-6">
-                      <v-col cols="12" md="4">
-                        <mew-button
-                          btn-size="small"
-                          btn-style="outline"
-                          title="Swap to renBTC"
-                          :has-full-width="true"
-                        />
-                      </v-col>
-                      <v-col cols="12" md="4">
-                        <mew-button
-                          btn-size="small"
-                          btn-style="outline"
-                          title="Swap to wBTC"
-                          :has-full-width="true"
-                        />
-                      </v-col>
-                      <v-col cols="12" md="4">
-                        <mew-button
-                          btn-size="small"
-                          btn-style="outline"
-                          title="Swap to PBTC"
-                          :has-full-width="true"
-                        />
-                      </v-col>
-                    </v-row>
-                  </v-expansion-panel-content>
-                </v-expansion-panel>
-              </v-expansion-panels>
-            </div>
-          </user-msg-block>
-
-          <!--
             =====================================================================================
               Address Book
             =====================================================================================
-            -->
+          -->
           <module-address-book
             v-show="showToAddress"
             class="mt-10"
             :is-valid-address-func="isValidToAddress"
             @setAddress="setToAddress"
           />
+
           <!--
             =====================================================================================
-             BTC options
+              User Message Block: store your Bitcoin on Ethereum
             =====================================================================================
-            -->
-          <div v-if="isToBtc" class="pa-6 wrapped-btc-text">
-            <p class="mew-heading-3">
-              <v-icon class="icon" size="20" color="#0B2840">
-                mdi-information-outline</v-icon
-              >Did you know? You can store your Bitcoin on Ethereum
-            </p>
-            <p>
-              To swap to BTC you need a Bitcoin wallet, but you can swap to
-              wrapped Bitcoin instead and store it in your Ethereum wallet.
-            </p>
-            <v-divider />
-            <v-expansion-panels flat>
-              <v-expansion-panel>
-                <v-expansion-panel-header color="transparent">
-                  How can I get wrapped Bitcoin?
-                </v-expansion-panel-header>
-                <v-expansion-panel-content color="transparent">
-                  <div>
-                    <p>
-                      When you swap to Bitcoin, it is moved to the Bitcoin
-                      blockchain, & requires a Bitcoin wallet. In order to keep
-                      Bitcoin in MyEtherWallet, you can swap to
-                      <span class="font-italic">wrapped</span> Bitcoin instead.
-                      Wrapped Bitcoin is an Ethereum token, with a value
-                      approximately equal to 1 BTC. Wrapped Bitcoins can be
-                      stored in MEW, and can be used as any other Ethereum
-                      asset: you can swap it to oether tokens, use it as
-                      collateral in DeFi apps, etc. There are multiple kinds of
-                      wrapped Bitcoins, but they roughly do the same thing.
-                      <a target="_blank" rel="noopener noreferrer"
-                        >Learn more about Wrapped Bitcoin.</a
-                      >
-                    </p>
-                    <div class="d-flex align-center">
-                      <mew-button
-                        v-for="btn in wrappedBtc"
-                        :key="btn"
-                        class="px-2 mx-1"
-                        :title="`Swap to ${btn}`"
-                        color-theme="primary"
-                        :has-full-width="false"
-                        btn-size="xlarge"
-                        btn-style="outline"
-                        @click.native="setWrappedBtc(btn)"
-                      />
-                    </div>
-                  </div>
-                </v-expansion-panel-content>
-              </v-expansion-panel>
-            </v-expansion-panels>
-          </div>
+          -->
+          <swap-msg-store-btc-eth
+            v-if="
+              toTokenType.value && toTokenType.value.toLowerCase() == 'bitcoin'
+            "
+            class="mt-5"
+            :message="msg.lowBalance"
+            :to-token-type="toTokenType.value"
+          />
+
           <!--
             =====================================================================================
              Providers List
             =====================================================================================
-            -->
+          -->
           <swap-providers-list
             :step="step"
             :available-quotes="availableQuotes"
@@ -285,7 +142,7 @@
 </template>
 
 <script>
-import UserMsgBlock from '@/views/components-wallet/TheWalletUserMsgBlock';
+import SwapMsgStoreBtcEth from './components/SwapMsgStoreBtcEth';
 import ModuleAddressBook from '@/modules/address-book/ModuleAddressBook';
 import SwapIcon from '@/assets/images/icons/icon-swap.svg';
 import SwapProvidersList from './components/SwapProvidersList.vue';
@@ -308,7 +165,7 @@ const MIN_GAS_WEI = '800000000000000';
 export default {
   name: 'ModuleSwap',
   components: {
-    UserMsgBlock,
+    SwapMsgStoreBtcEth,
     ModuleAddressBook,
     SwapProvidersList,
     SwapFee,