+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Devop
 
+* Fix Swap store BTC on ETH message [#3106](https://github.com/MyEtherWallet/MyEtherWallet/pull/3106)
 * Fix mew-select border issue [#3091](https://github.com/MyEtherWallet/MyEtherWallet/pull/3091)
 * Add user msg for swap [#3087](https://github.com/MyEtherWallet/MyEtherWallet/pull/3087)
 * Fix ui issues for aave [#3079](https://github.com/MyEtherWallet/MyEtherWallet/pull/3079)